Fix: 2 int to 2 float

+++ Backend/schemas.py
@@ -16,8 +16,8 @@ class Item(ItemBase):
 
 class ActividadBase(BaseModel):
     nombre: str
-    temperatura_min: int
-    temperatura_max: int
+    temperatura_min: float
+    temperatura_max: float
     estado_dia: str
     descripcion: str
